refactor: extract reusable workbook components and fix card sizing
- Extract reusable components to app/workbook/_components/:
  - WorkbookAuthGuard: authentication wrapper
  - WorkbookHeader: reusable header with back button
  - WorkbookDataProvider: data fetching and state management
  - WorkbookContent: content display with loading states
  - WorkbookInfiniteScroll: infinite scroll logic
- Refactor practice page to use new components (375 → 125 lines)
- Fix TestCard width issues: remove desktop:w-6xl, use desktop:w-[30vw]
- Remove font-['Inter'] classes for cleaner styling
- Maintain pathname-based URL handling for correct routing
