+++ app/_components/cards/TestCard.tsx
@@ -10,7 +10,7 @@ interface TestCardProps {
 export default function TestCard({ solves, category }: TestCardProps) {
   if (!solves?.length) {
     return (
-      <div className="inline-flex w-96 flex-col items-center justify-center overflow-hidden rounded-2xl bg-white/90 p-4 shadow-lg outline-1 outline-offset-[-1px] outline-zinc-300 backdrop-blur-[2px] desktop:min-w-96 desktop:grow">
+      <div className="inline-flex w-96 flex-col items-center justify-center overflow-hidden rounded-2xl bg-white/90 p-4 shadow-lg outline-1 outline-offset-[-1px] outline-zinc-300 backdrop-blur-[2px] desktop:w-[480px]">
         <div className="text-sm text-gray-600">데이터가 없습니다</div>
       </div>
     );
@@ -36,15 +36,15 @@ export default function TestCard({ solves, category }: TestCardProps) {
     : '';
 
   return (
-    <div className="inline-flex w-96 flex-col items-start justify-start overflow-hidden rounded-2xl bg-white/90 p-4 shadow-lg outline-1 outline-offset-[-1px] outline-zinc-300 backdrop-blur-[2px] desktop:w-6xl desktop:min-w-96 desktop:grow">
+    <div className="inline-flex w-96 flex-col items-start justify-start overflow-hidden rounded-2xl bg-white/90 p-4 shadow-lg outline-1 outline-offset-[-1px] outline-zinc-300 backdrop-blur-[2px] desktop:w-[30vw]">
       <div className="flex flex-col items-start justify-start self-stretch">
         <div className="inline-flex items-center justify-start self-stretch">
-          <div className="justify-end font-['Inter'] text-lg leading-7 font-bold text-gray-800">
+          <div className="justify-end text-lg leading-7 font-bold text-gray-800">
             {category}
           </div>
         </div>
         <div className="inline-flex items-end justify-end self-stretch">
-          <div className="justify-center font-['Inter'] text-xs leading-7 font-normal text-black">
+          <div className="justify-center text-xs leading-7 font-normal text-black">
             {totalProblems} 문제
           </div>
         </div>
@@ -55,7 +55,7 @@ export default function TestCard({ solves, category }: TestCardProps) {
             key={solve.id}
             className="inline-flex items-center justify-start self-stretch rounded bg-gray-50 p-2"
           >
-            <div className="justify-center font-['Inter'] text-sm leading-tight font-normal text-gray-700">
+            <div className="justify-center text-sm leading-tight font-normal text-gray-700">
               {solve.question}
             </div>
           </div>
@@ -63,11 +63,11 @@ export default function TestCard({ solves, category }: TestCardProps) {
       </div>
       <div className="mt-auto flex flex-col items-start justify-start gap-1 self-stretch pt-1">
         <div className="inline-flex items-start justify-between self-stretch">
-          <div className="justify-center font-['Inter'] text-sm leading-tight font-normal text-gray-600">
+          <div className="justify-center text-sm leading-tight font-normal text-gray-600">
             정답률
           </div>
           <div className="inline-flex flex-col items-start justify-start self-stretch">
-            <div className="justify-center font-['Inter'] text-sm leading-tight font-normal text-gray-600">
+            <div className="justify-center text-sm leading-tight font-normal text-gray-600">
               {accuracy}%
             </div>
           </div>
@@ -80,7 +80,7 @@ export default function TestCard({ solves, category }: TestCardProps) {
         </div>
       </div>
       <div className="flex flex-col items-start justify-end gap-2.5 self-stretch pt-3">
-        <div className="justify-center font-['Inter'] text-sm leading-tight font-normal text-gray-600">
+        <div className="justify-center text-sm leading-tight font-normal text-gray-600">
           {latestDate}
         </div>
       </div>
